IGNITE-28976 [ducktests] Split the demo breakpoint checks into focused files and shared helpers
check_pause.py held four unrelated concerns plus a set of ducktape test
doubles that are not pause specific at all. Split it by concern into
check_pause_selector, check_pause_control, check_pause_timeout and
check_pause_banner, and move the reusable parts out of checks/utils - which
mirrors ignitetest/utils and is meant for check files alone - into
checks/support, where any check can pick them up.

checks/ carries no __init__.py on purpose, so setup.py's find_packages()
keeps it out of the distribution; a conftest.py at the tests root is what
puts that root on sys.path and lets checks.support resolve as a namespace
package.

Same checks, same assertions, no framework code touched.
